+++ resources/[dg]/dg-api/typescript/src/server/types/api.d.ts
@@ -1,5 +1,5 @@
 declare namespace API {
-  type Method = 'GET' | 'POST' | 'PUT' | 'DELETE';
+  type Method = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH';
 
   type Responser = (code: number, data: any, headers?: Record<string, string>) => void;