+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - Developer optie toegevoegd om item metadata aan te passen
+- API endpoint om account balance te veranderen
 
 ### Changed